+++ src/packages/documents/documents/rollback/entity-action/manifests.ts
@@ -1,5 +1,6 @@
 import { UMB_DOCUMENT_ENTITY_TYPE } from '../../entity.js';
 import { UMB_USER_PERMISSION_DOCUMENT_ROLLBACK } from '../../user-permissions/index.js';
+import { UMB_DOCUMENT_WORKSPACE_ALIAS } from '../../workspace/index.js';
 import { UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
@@ -23,6 +24,12 @@ export const manifests: Array<ManifestTypes> = [
 			},
 			{
 				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
+			},
+			/* Currently the rollback is tightly coupled to the workspace contexts so we only allow it to show up
+			 In the document workspace. */
+			{
+				alias: 'Umb.Condition.WorkspaceAlias',
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 		],
 	},
