+++ src/packages/documents/documents/workspace/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DOCUMENT_WORKSPACE_ALIAS = 'Umb.Workspace.Document';
