+++ src/packages/documents/documents/workspace/document-workspace.context.ts
@@ -26,7 +26,7 @@ import {
 import { UMB_DOCUMENTS_SECTION_PATH } from '../../section/paths.js';
 import { UmbDocumentPreviewRepository } from '../repository/preview/index.js';
 import { sortVariants } from '../utils.js';
-import { UMB_DOCUMENT_WORKSPACE_ALIAS } from './manifests.js';
+import { UMB_DOCUMENT_WORKSPACE_ALIAS } from './constants.js';
 import { UmbEntityContext } from '@umbraco-cms/backoffice/entity';
 import { UMB_INVARIANT_CULTURE, UmbVariantId } from '@umbraco-cms/backoffice/variant';
 import { UmbContentTypeStructureManager } from '@umbraco-cms/backoffice/content-type';
