+++ src/packages/documents/documents/workspace/index.ts
@@ -1,2 +1,3 @@
 export * from './document-workspace.context-token.js';
 export * from './document-workspace.context.js';
+export * from './constants.js';
