+++ src/packages/documents/documents/workspace/manifests.ts
@@ -4,6 +4,7 @@ import {
 	UMB_USER_PERMISSION_DOCUMENT_PUBLISH,
 } from '../user-permissions/index.js';
 import { UMB_DOCUMENT_ENTITY_TYPE } from '../entity.js';
+import { UMB_DOCUMENT_WORKSPACE_ALIAS } from './constants.js';
 import { UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
 import type {
 	ManifestWorkspaces,
@@ -13,8 +14,6 @@ import type {
 	ManifestTypes,
 } from '@umbraco-cms/backoffice/extension-registry';
 import { UMB_CONTENT_HAS_PROPERTIES_WORKSPACE_CONDITION } from '@umbraco-cms/backoffice/content';
-
-export const UMB_DOCUMENT_WORKSPACE_ALIAS = 'Umb.Workspace.Document';
 
 const workspace: ManifestWorkspaces = {
 	type: 'workspace',
@@ -41,7 +40,7 @@ const workspaceViews: Array<ManifestWorkspaceView> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 			{
 				alias: 'Umb.Condition.WorkspaceHasCollection',
@@ -62,7 +61,7 @@ const workspaceViews: Array<ManifestWorkspaceView> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 			{
 				alias: UMB_CONTENT_HAS_PROPERTIES_WORKSPACE_CONDITION,
@@ -83,7 +82,7 @@ const workspaceViews: Array<ManifestWorkspaceView> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 		],
 	},
@@ -105,7 +104,7 @@ const workspaceActions: Array<ManifestWorkspaceActions> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 			{
 				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
@@ -127,7 +126,7 @@ const workspaceActions: Array<ManifestWorkspaceActions> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 			{
 				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
@@ -147,7 +146,7 @@ const workspaceActions: Array<ManifestWorkspaceActions> = [
 		conditions: [
 			{
 				alias: 'Umb.Condition.WorkspaceAlias',
-				match: workspace.alias,
+				match: UMB_DOCUMENT_WORKSPACE_ALIAS,
 			},
 			{
 				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
