+++ src/packages/documents/documents/workspace/views/info/document-workspace-view-info-history.element.ts
@@ -66,16 +66,6 @@ export class UmbDocumentWorkspaceViewInfoHistoryElement extends UmbLitElement {
 		this.#requestAuditLogs();
 	}
 
-	#onRollbackModalOpen = async () => {
-		const modalManagerContext = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
-		const modalContext = modalManagerContext.open(this, UMB_ROLLBACK_MODAL, {});
-
-		await modalContext.onSubmit();
-		// TODO: This notification won't actually show at the moment because we perform a full page reload after rollback. However, when we can do it without a full page reload, this should be used.
-		const notificationContext = await this.getContext(UMB_NOTIFICATION_CONTEXT);
-		notificationContext.peek('positive', { data: { message: this.localize.term('rollback_documentRolledBack') } });
-	};
-
 	async #requestAndCacheUserItems() {
 		const allUsers = this._items?.map((item) => item.user.unique).filter(Boolean) as string[];
 		const uniqueUsers = [...new Set(allUsers)];
@@ -98,12 +88,10 @@ export class UmbDocumentWorkspaceViewInfoHistoryElement extends UmbLitElement {
 	override render() {
 		return html`<uui-box>
 			<umb-localize slot="headline" key="general_history">History</umb-localize>
-			<uui-button
+			<umb-extension-with-api-slot
 				slot="header-actions"
-				label=${this.localize.term('actions_rollback')}
-				look="secondary"
-				@click=${this.#onRollbackModalOpen}>
-				<uui-icon name="icon-undo"></uui-icon> ${this.localize.term('actions_rollback')}
+				type="entityAction"
+				.filter=${(manifest: any) => manifest.alias === 'Umb.EntityAction.Document.Rollback'}></umb-extension-with-api-slot>
 			</uui-button>
 			${this._items ? this.#renderHistory() : html`<uui-loader-circle></uui-loader-circle> `}
 			${this.#renderPagination()}
